fixed early ending bug
Playtester put bird on the bird slot before putting the mouse in the kitchen, and ended the story early. I was debating keeping it in, as like an alternate ending but it wouldn't really make sense since the mouse has to eat somehow, and she doesn't exactly know how to do it right.